requirements.txt
# ── Core ML ───────────────────────────────────────────────────────────────────
mlflow>=2.13.0
xgboost>=2.0.0
scikit-learn>=1.4.0

# ── Data handling ─────────────────────────────────────────────────────────────
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=15.0.0          # Parquet support for pandas

# ── Drift detection ───────────────────────────────────────────────────────────
evidently>=0.4.0

# ── Model serving ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
prometheus-fastapi-instrumentator>=7.0.0

# ── Configuration & environment ───────────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.1

# ── HTTP client (used by MLflow & health checks) ──────────────────────────────
httpx>=0.27.0
requests>=2.31.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0            # async test client for FastAPI

# ── Dev / linting (optional — not installed in Docker) ───────────────────────
# flake8>=7.0.0
# black>=24.0.0
# isort>=5.13.0
